Use bool return value for RequestBlock when none possible
This fixes a bug in the request manager when we try to request two
thinblocks at the same time from the same peer.  We can not currently
handle to thinkblocks from the same peer concurrently and so what was
happening was we would not request the thinblock but the request
manager was removing the block source.  By using a bool return we
can check to see if the thinkblock was actually requested and then
update the block source accordingly.  This but would cause frequent
hangups in during regression testing and was another cause of
blocks being re-requested.

- Remove regtest workaround for retry interval

With the above fix for requesting thinblocks we no longer
need the regtest workaround for requesting blocks.