Add IsThinBlockValid() when processing thinblock or xthinblock
Check whether there are any transactions in the thinblock and
whether the first transaction is a coinbase.

Add CheckBlockHeader() to IsThinBlockValid()

Remove dead code for CheckBlockHeader in CXThinblock

Now that we have IsThinBlockValid() we no longer need nor were
using CheckBlockHeader in CXThinblock.

Add thinblock test for invalid block header