Ensure that we connect to at least a few XTHIN capable nodes.
In the past we relied on -addnode or -connect-thinblock to manually
configure connection to XTHIN capable nodes,  or to just rely on luck
in getting a connection.  Here we will actively seek out XTHIN nodes
and maintain at least 4 connections or the minimum number of outbound
connections whichever is less.