When searching for XTHIN capable nodes, check whether IBD is complete.
During IBD we do not check and search for XTHIN capable nodes
however by the time IBD is completed all the outgoing connections are
already in use and the netcode is blocked waiting for a semaphore grant.
Therefore we need to use a try_wai() on the semaphore and then checked
every so often whether IBD is complete and if so then begin the search
for XTHIN nodes.